Remove the obsolete reader.* and rename writer.* to readerwriter.*.
