+++ arch/apple2/encoder.cc
@@ -3,7 +3,7 @@
 #include "decoders/decoders.h"
 #include "encoders/encoders.h"
 #include "sector.h"
-#include "writer.h"
+#include "readerwriter.h"
 #include "image.h"
 #include "fmt/format.h"
 #include "lib/encoders/encoders.pb.h"
