+++ arch/tids990/encoder.cc
@@ -3,7 +3,7 @@
 #include "encoders/encoders.h"
 #include "tids990.h"
 #include "crc.h"
-#include "writer.h"
+#include "readerwriter.h"
 #include "image.h"
 #include "arch/tids990/tids990.pb.h"
 #include "lib/encoders/encoders.pb.h"
