+++ src/fe-rawwrite.cc
@@ -1,8 +1,7 @@
 #include "globals.h"
 #include "flags.h"
-#include "reader.h"
+#include "readerwriter.h"
 #include "fluxmap.h"
-#include "writer.h"
 #include "sector.h"
 #include "proto.h"
 #include "lib/fluxsource/fluxsource.h"
