+++ src/gui/mainwindow.cc
@@ -2,8 +2,7 @@
 #include "proto.h"
 #include "gui.h"
 #include "logger.h"
-#include "reader.h"
-#include "writer.h"
+#include "readerwriter.h"
 #include "fluxsource/fluxsource.h"
 #include "fluxsink/fluxsink.h"
 #include "imagereader/imagereader.h"
